Move `use_given_version()` to `use_command` module
