+++ src/pycors.rs
@@ -1,11 +1,9 @@
-use failure::format_err;
 use prettytable::{cell, row, Cell, Row, Table};
-use semver::VersionReq;
 use structopt::StructOpt;
 
 use crate::commands;
 use crate::config::Cfg;
-use crate::settings::{PythonVersion, Settings};
+use crate::settings::Settings;
 use crate::utils;
 use crate::Result;
 use crate::{commands::Command, Opt};
@@ -22,7 +20,9 @@ pub fn pycors(cfg: &Option<Cfg>, settings: &Settings) -> Result<()> {
             Command::List => print_to_stdout_available_python_versions(cfg, settings)?,
             Command::Path => print_active_interpreter_path(cfg, settings)?,
             Command::Version => print_active_interpreter_version(cfg, settings)?,
-            Command::Use { version } => use_given_version(&version, settings)?,
+            Command::Use { version } => {
+                commands::use_command::use_given_version(&version, settings)?
+            }
             Command::Install { from_version } => {
                 commands::install::install_python(from_version, cfg, settings)?;
             }
@@ -44,43 +44,6 @@ fn print_active_interpreter_path(cfg: &Option<Cfg>, settings: &Settings) -> Resu
 fn print_active_interpreter_version(cfg: &Option<Cfg>, settings: &Settings) -> Result<()> {
     let interpreter_to_use = utils::get_interpreter_to_use(cfg, settings)?;
     println!("{}", interpreter_to_use.version);
-    Ok(())
-}
-
-fn use_given_version(requested_version: &str, settings: &Settings) -> Result<()> {
-    // Convert the requested version string to proper VersionReq
-    // FIXME: Should a `~` be explicitly added here if user does not provide it?
-    log::debug!("Requesting version: {}", requested_version);
-    let version: VersionReq = requested_version.parse()?;
-    log::debug!("Semantic version requirement: {}", version);
-
-    let python_to_use = match utils::active_version(&version, settings) {
-        Some(python_to_use) => python_to_use.clone(),
-        None => {
-            let new_cfg = Some(Cfg { version });
-            let version = commands::install::install_python(None, &new_cfg, settings)?
-                .ok_or_else(|| format_err!("A Python version should have been installed"))?;
-            let install_dir = utils::install_dir(&version)?;
-
-            PythonVersion {
-                version,
-                location: install_dir,
-            }
-        }
-    };
-
-    log::debug!(
-        "Using {} from {}",
-        python_to_use.version,
-        python_to_use.location.display()
-    );
-
-    // Write to `.python-version`
-    Cfg {
-        version: VersionReq::exact(&python_to_use.version),
-    }
-    .save()?;
-
     Ok(())
 }
